Add support for numeric UIDs/GIDs for files and directories
There are cases where it's necessary to use chown with a UID/GID instead
of a username because the ID of the owner/group isn't present in /etc/passwd.
This could be because the file is inside a NFS-shared directory or
because it's bind-mounted into a (uidmapped) container environment.

Currently, when using a UID instead of a username, the chown works because
it recognizes the value as ID but then afterwards the verify fails to
verify its action because it compares the output of `stat -c '%U'` with
the `owner` attribute of the File item and `stat` outputs `UNKNOWN` for
UIDs/GIDs it can't map back to a username.

To solve this, support for letting files be owned by a UID/GID instead of
a username is added in form of UIDs/GIDs prefixed by a `+` and if so,
use `stat -c '%u'` for validation to check against the UID instead of the
username.
The +-Syntax is directly taken from `chown` (both BSD and GNU) and therefore
the fix-part just works and only the sdict/check function needs to be adjusted.